Add proper widget implementation for window list
