+++ src/daemon/gui/mod.rs
@@ -1,8 +1,6 @@
 /* niri-switch  Copyright (C) 2025  Kiki/Bouba Team */
 mod store;
 mod style;
-mod window_info;
-mod window_item;
 mod window_list;
 
 use super::dbus;
@@ -16,6 +14,7 @@ use std::{
     collections::{HashMap, HashSet},
     sync::{Arc, Mutex},
 };
+use window_list::WindowList;
 
 /* Type aliases to make signatures more readable */
 type GlobalStoreRef = Arc<Mutex<store::GlobalStore>>;
@@ -62,7 +61,7 @@ fn sort_windows_by_cached_order(windows: &mut [niri_ipc::Window], store: &Global
 }
 
 /// Handle request to activate the daemon
-async fn handle_daemon_activated(list: &gtk4::ListView, store: &GlobalStoreRef) {
+async fn handle_daemon_activated(list: &WindowList, store: &GlobalStoreRef) {
     let window = list
         .root()
         .and_downcast::<gtk4::Window>()
@@ -70,12 +69,12 @@ async fn handle_daemon_activated(list: &gtk4::ListView, store: &GlobalStoreRef) 
 
     /* If window is already shown, simply advance the selection */
     if window.is_visible() {
-        window_list::advance_the_selection(list);
+        list.advance_the_selection();
         return;
     }
     /* Else reload the listed windows, state might have changed since the last time.
      * This is also the initial filling of the list. */
-    window_list::clear_the_list(list);
+    list.clear_the_list();
 
     /* niri socket uses blocking calls, so it will be run on a separate thread */
     let store_ref = store.clone();
@@ -103,17 +102,17 @@ async fn handle_daemon_activated(list: &gtk4::ListView, store: &GlobalStoreRef) 
     }
 
     /* Append windows to the list model */
-    window_list::fill_the_list(list, &windows, store);
+    list.fill_the_list(&windows, store);
 
     /* Next bring the window back to visibility */
     window.present();
 
     /* List will loose focus after droping the elements, need to grab it again */
-    list.grab_focus();
+    list.focus_to_list();
 }
 
 /// Handle event from the D-Bus connection
-async fn handle_dbus_event(event: dbus::DbusEvent, list: &gtk4::ListView, store: &GlobalStoreRef) {
+async fn handle_dbus_event(event: dbus::DbusEvent, list: &WindowList, store: &GlobalStoreRef) {
     use dbus::DbusEvent::*;
     match event {
         Activate => handle_daemon_activated(list, store).await,
@@ -123,12 +122,21 @@ async fn handle_dbus_event(event: dbus::DbusEvent, list: &gtk4::ListView, store:
 /// Creates the main window, widgets, models and factories
 fn activate(application: &gtk4::Application, global_store: &GlobalStoreRef) {
     /* Create widget for displaying windows */
-    let list_view = window_list::create_window_list(global_store);
+    let window_list = window_list::WindowList::default();
+
+    /* clone! macro will create another reference to the store object, so it can be moved
+     * to the closure. The closure can outlive ther current function scope, so it has hold
+     * own reference. */
+    window_list.connect_activate(clone!(
+        #[strong]
+        global_store,
+        move |list, position| window_list::handle_window_chosen(list, position, &global_store)
+    ));
 
     /* Create main window */
     let window = gtk4::ApplicationWindow::builder()
         .application(application)
-        .child(&list_view)
+        .child(&window_list)
         .build();
 
     /* Create a weak reference to the window, this will be moved to keyboard controller
@@ -161,12 +169,12 @@ fn activate(application: &gtk4::Application, global_store: &GlobalStoreRef) {
     /* Start a task that handles events from D-Bus */
     glib::spawn_future_local(clone!(
         #[weak]
-        list_view,
+        window_list,
         #[strong]
         global_store,
         async move {
             while let Ok(event) = receiver.recv().await {
-                handle_dbus_event(event, &list_view, &global_store).await;
+                handle_dbus_event(event, &window_list, &global_store).await;
             }
         }
     ));
