+++ src/daemon/gui/window_list/window_info/imp.rs
