+++ src/daemon/gui/window_list/window_info/mod.rs
