+++ src/daemon/gui/window_list/window_item/window_item.ui
